Rework structure of Controller
