--- src/LMSController/ClientInformation.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace LMSController
-{
-    public class ClientInformation
-    {
-        public int LibraryTicketNumberCode { get; set; }
-        public PersonalInformation PersonalInformation { get; set; }
-    }
-}
