--- src/LMSController/UserCredential.cs
@@ -1,8 +1,0 @@
-﻿namespace LMSController
-{
-    public class UserCredential
-    {
-        public string Name { get; set; }
-        public string Password { get; set; }
-    }
-}
