+++ src/LMSController/administrator/AdminWorkspace.cs
@@ -2,7 +2,7 @@
 {
    internal class AdminWorkspace : IWorkspaceAdmin
    {
-      private static readonly IUserInformationRegister userInformationRegister = new UserInformationRegister();
+      private static readonly UserInformationRegister userInformationRegister = new UserInformationRegister();
       public IUserInformationRegister UserInformationRegister => userInformationRegister;
    }
 }
