+++ src/LMSController/administrator/IUserInformationRegister.cs
