+++ src/LMSController/authentication/IAuthentication.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace LMSController
+﻿namespace LMSController
 {
    public delegate void LibrarianLogOnEventHandler(IWorkspaceLibrarian workspaceLibrarian);
    public delegate void AdminLogOnEventHandler(IWorkspaceAdmin workspaceLibrarian);
