+++ src/LMSController/doc/LMSControllerAdministrator.cd
@@ -3,37 +3,34 @@
   <Class Name="LMSController.AdminWorkspace" BaseTypeListCollapsed="true">
     <Position X="4.25" Y="0.5" Width="1.5" />
     <Members>
-      <Field Name="userInformationRegister" Hidden="true" />
+      <Property Name="UserInformationRegister" Hidden="true" />
     </Members>
+    <AssociationLine Name="userInformationRegister" Type="LMSController.UserInformationRegister">
+      <MemberNameLabel ManuallyPlaced="true">
+        <Position X="0.1" Y="0.3" />
+      </MemberNameLabel>
+    </AssociationLine>
     <TypeIdentifier>
       <HashCode>AAAAAAAAAAAAAAAAAQAAAAAAAAAAAAAAAQAAAAAAAAA=</HashCode>
-      <FileName>AdminWorkspace.cs</FileName>
+      <FileName>administrator\AdminWorkspace.cs</FileName>
     </TypeIdentifier>
     <ShowAsAssociation>
       <Field Name="userInformationRegister" />
-      <Property Name="UserInformationRegister" />
     </ShowAsAssociation>
     <Lollipop Position="0.2" />
   </Class>
-  <Class Name="LMSController.PersonalInformation">
-    <Position X="7.5" Y="5" Width="2" />
-    <TypeIdentifier>
-      <HashCode>AAAAAAAAAAAAAIAAAAAAAAAAAAAIAAACAAAAIAAAAAA=</HashCode>
-      <FileName>PersonalInformation.cs</FileName>
-    </TypeIdentifier>
-  </Class>
   <Class Name="LMSController.UserCredential">
-    <Position X="0.5" Y="5" Width="2" />
+    <Position X="0.5" Y="4.25" Width="2" />
     <TypeIdentifier>
       <HashCode>AAAAAAAAAAAAAAAAAAAAAAQAAAAAAAACAAAAAAAAAAA=</HashCode>
-      <FileName>UserCredential.cs</FileName>
+      <FileName>types\UserCredential.cs</FileName>
     </TypeIdentifier>
   </Class>
   <Class Name="LMSController.UserInformation">
-    <Position X="4" Y="5" Width="2" />
+    <Position X="4" Y="4.25" Width="2" />
     <TypeIdentifier>
       <HashCode>AAAAAAAAAAAAAAAgAAAAAAgAAAAAAAAAAAAABAIAAAA=</HashCode>
-      <FileName>UserInformation.cs</FileName>
+      <FileName>types\UserInformation.cs</FileName>
     </TypeIdentifier>
     <ShowAsAssociation>
       <Property Name="Credential" />
@@ -41,19 +38,19 @@
     </ShowAsAssociation>
   </Class>
   <Class Name="LMSController.UserInformationRegister" BaseTypeListCollapsed="true">
-    <Position X="4" Y="2.75" Width="2" />
+    <Position X="4" Y="2" Width="2" />
     <TypeIdentifier>
       <HashCode>gAAAAAAAAAAAAAAEAAAAAAAEAAAAAAAgAAAAAAAAAgA=</HashCode>
-      <FileName>UserInformationRegister.cs</FileName>
+      <FileName>administrator\UserInformationRegister.cs</FileName>
     </TypeIdentifier>
     <Lollipop Position="0.2" />
   </Class>
-  <Interface Name="LMSController.IUserInformationRegister" Collapsed="true">
-    <Position X="4" Y="1.75" Width="2" />
+  <Struct Name="LMSController.PersonalInformation">
+    <Position X="8" Y="4.25" Width="1.5" />
     <TypeIdentifier>
-      <HashCode>gAAAAAAAAAAAAAAEAAAAAAAEAAAAAAAgAAAAAAAAAgA=</HashCode>
-      <FileName>IUserInformationRegister.cs</FileName>
+      <HashCode>AAAAAAAAAAAAAIAAAAAAAAAAAAAIAAACAAAAIAAAAAA=</HashCode>
+      <FileName>types\PersonalInformation.cs</FileName>
     </TypeIdentifier>
-  </Interface>
+  </Struct>
   <Font Name="Segoe UI" Size="9" />
 </ClassDiagram>
