+++ src/LMSController/stockman/IPublishHouseInformationRegister.cs
