+++ src/LMSController/stockman/ISubjectInformationFinding.cs
@@ -1,0 +1,7 @@
+﻿namespace LMSController
+{
+   public interface ISubjectInformationFinding
+   {
+      SubjectInformation FindByName(string name);
+   }
+}
