+++ src/LMSController/stockman/StockmanWorkspace.cs
@@ -2,8 +2,8 @@
 {
    internal class StockmanWorkspace : IWorkspaceStockman
    {
-      private static readonly IBookFinding bookFinding = new BookFinding();
-      private static readonly ILibraryOperations libraryOperations = new LibraryOperations();
+      private static readonly BookFinding bookFinding = new BookFinding();
+      private static readonly LibraryOperations libraryOperations = new LibraryOperations();
 
       public IBookFinding BookFinding => bookFinding;
       public ILibraryOperations LibraryOperations => libraryOperations;
