+++ src/LMSController/types/AuthorInformation.cs
@@ -1,0 +1,7 @@
+﻿namespace LMSController
+{
+   public struct AuthorInformation
+   {
+      public string Name { get; set; }
+   }
+}
