+++ src/LMSController/types/ClientInformation.cs
@@ -1,0 +1,8 @@
+﻿namespace LMSController
+{
+   public class ClientInformation
+   {
+      public int LibraryTicketNumberCode { get; set; }
+      public PersonalInformation PersonalInformation { get; set; }
+   }
+}
