+++ src/LMSController/types/PersonalInformation.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace LMSController
+{
+   public struct PersonalInformation
+   {
+      public string FullName { get; set; }
+      public DateTime Birthday { get; set; }
+      public decimal Phone { get; set; }
+      public string Address { get; set; }
+   }
+}
