+++ src/LMSController/types/PublishHouseInformation.cs
@@ -1,0 +1,9 @@
+﻿namespace LMSController
+{
+   public class PublishHouseInformation
+   {
+      public string Name { get; set; }
+      public string Location { get; set; }
+      public string Category { get; set; }
+   }
+}
