+++ src/LMSController/types/SubjectInformation.cs
@@ -1,0 +1,8 @@
+﻿namespace LMSController
+{
+   public class SubjectInformation
+   {
+      public string Name { get; set; }
+      public SubjectInformation SubjectParent { get; set; }
+   }
+}
