+++ src/LMSController/types/UserCredential.cs
@@ -1,7 +1,8 @@
 ﻿namespace LMSController
 {
-   public class AuthorInformation
+   public class UserCredential
    {
       public string Name { get; set; }
+      public string Password { get; set; }
    }
 }
