+++ src/LMSController/types/UserInformation.cs
@@ -1,0 +1,11 @@
+﻿namespace LMSController
+{
+   public class UserInformation
+   {
+      internal int UserId { get; set; }
+      
+      public UserCredential Credential { get; set; }
+      public string Position { get; set; }
+      public PersonalInformation PersonalInformation { get; set; }
+   }
+}
